Documentation for events endpoints. Consolidated events endpoints to remove redundant endpoints.
