--- client/src/api/traffic/deleteSeries.js
@@ -1,6 +1,0 @@
-import axios from 'axios';
-import { ROOT_TRAFFIC_URL } from '../root';
-
-export function deleteSeries(trafficSeriesId) {
-  return axios.delete(ROOT_TRAFFIC_URL + '/' + trafficSeriesId);
-}
