+++ client/src/api/traffic/index.js
@@ -1,25 +1,23 @@
 import { add } from './add';
 import { createInstance } from './createInstance';
-import { deleteInstance } from './deleteInstance';
-import { deleteSeries } from './deleteSeries';
 import { find } from './find';
 import { findById } from './findById';
 import { findEarliest } from './findEarliest';
+import { remove } from './remove';
+import { removeInstanceFromSeries } from './removeInstanceFromSeries';
 import { search } from './search';
 import { searchUnderwriters } from './searchUnderwriters';
-import { updateInstance } from './updateInstance';
-import { updateSeries } from './updateSeries';
+import { update } from './update';
 
 export const trafficAPI = {
   add,
   createInstance,
-  deleteInstance,
-  deleteSeries,
   find,
   findById,
   findEarliest,
+  remove,
+  removeInstanceFromSeries,
   search,
   searchUnderwriters,
-  updateInstance,
-  updateSeries,
+  update,
 };
