+++ client/src/api/traffic/remove.js
@@ -1,0 +1,6 @@
+import axios from 'axios';
+import { ROOT_TRAFFIC_URL } from '../root';
+
+export function remove(trafficId) {
+  return axios.delete(ROOT_TRAFFIC_URL + '/' + trafficId);
+}
