+++ client/src/redux/traffic/actions/index.js
@@ -1,29 +1,27 @@
 import { add } from './add';
 import { clear } from './clear';
 import { createInstance } from './createInstance';
-import { deleteInstance } from './deleteInstance';
-import { deleteSeries } from './deleteSeries';
 import { find } from './find';
 import { findEarliest } from './findEarliest';
 import { findById } from './findById';
+import { remove } from './remove';
+import { removeInstanceFromSeries } from './removeInstanceFromSeries';
 import { search } from './search';
 import { searchUnderwriters } from './searchUnderwriters';
 import { setReturnPageAfterAddDeleteActions } from './setReturnPageAfterAddDeleteActions';
-import { updateInstance } from './updateInstance';
-import { updateSeries } from './updateSeries';
+import { update } from './update';
 
 export const trafficActions = {
   add,
   clear,
   createInstance,
-  deleteInstance,
-  deleteSeries,
   find,
   findById,
   findEarliest,
+  remove,
+  removeInstanceFromSeries,
   search,
   searchUnderwriters,
   setReturnPageAfterAddDeleteActions,
-  updateInstance,
-  updateSeries,
+  update,
 };
