+++ client/src/redux/traffic/actions/removeInstanceFromSeries.js
@@ -1,9 +1,15 @@
 import { alertTypes } from '../../alert';
 import { trafficAPI } from '../../../api';
 
-export const deleteSeries = (trafficSeriesId, callback) => async dispatch => {
+export const removeInstanceFromSeries = (
+  traffic,
+  callback,
+) => async dispatch => {
   try {
-    const response = await trafficAPI.deleteSeries(trafficSeriesId);
+    const response = await trafficAPI.removeInstanceFromSeries(
+      traffic.master_event_id._id,
+      traffic,
+    );
     callback(response);
   } catch (e) {
     console.error(e);
