+++ client/src/redux/traffic/actions/update.js
@@ -1,9 +1,13 @@
 import { alertTypes } from '../../alert';
 import { trafficAPI } from '../../../api';
 
-export const updateSeries = (data, callback) => async dispatch => {
+export const update = (
+  instanceId,
+  changedValues,
+  callback,
+) => async dispatch => {
   try {
-    const response = await trafficAPI.updateSeries(data._id, data);
+    const response = await trafficAPI.update(instanceId, changedValues);
     callback(response);
   } catch (e) {
     console.error(e);
