--- server/v1/controllers/events/instance/index.js
@@ -1,5 +1,0 @@
-const remove = require('./remove');
-
-module.exports = {
-  remove,
-};
