+++ server/v1/controllers/events/nextShow.js
@@ -1,22 +1,22 @@
 /**
  * @swagger
  *
- * /current-show:
+ * /next-show:
  *   get:
  *     tags:
  *     - Simple Endpoints
  *     - Shows
- *     operationId: CurrentShow
- *     summary: Current Show
+ *     operationId: NextShow
+ *     summary: Next Show
  *     security:
  *     - ApiKeyAuth: []
  *     description: |
- *       Returns the show that's currently playing, or null if there is no show currently playing.
+ *       Returns the next show that will be playing. Returns `null` if there are no shows occurring within the next day.
  *
  *       The following roles can access this API endpoint: `Admin`, `Full Access`, `Show Captain`, `Underwriting`, `DJ`, `Music Library Admin`
  *     responses:
  *       200:
- *         description: The show that is currently playing, including its playlist. `null` if there is no show currently playing.
+ *         description: An object containing the next show, including its playlist. `null` if no shows occur during the next day.
  *         content:
  *           application/json:
  *             schema:
@@ -171,8 +171,8 @@ const {
     populateMasterEvent,
     populateMasterEventShowDetails,
   },
-} = require('../utils');
-const { findOrCreatePlaylist } = require('../../playlists/utils');
+} = require('./utils');
+const { findOrCreatePlaylist } = require('../playlists/utils');
 
 function currentShow(req, res) {
   const dbModel = getModelForEventType('shows');
@@ -182,14 +182,23 @@ function currentShow(req, res) {
   }
 
   let startDate = new Date();
-  let endDate = new Date(startDate.getTime() + 60 * 60 * 1000);
+  let endDate = new Date(startDate.getTime() + 24 * 60 * 60 * 1000);
 
   let filter = findEventQueryByDateRange(startDate, endDate);
 
   const processEventResults = dbShow => {
     let showResults = eventList(dbShow, startDate, endDate);
     if (showResults.length > 0) {
-      let show = showResults[0];
+      let show;
+      for (let i = 0; i < showResults.length; i++) {
+        if (new Date(showResults[i].start_time_utc) > startDate) {
+          show = showResults[i];
+          break;
+        }
+      }
+      if (typeof show === 'undefined') {
+        return res.json(null);
+      }
       return findOrCreatePlaylist(show.start_time_utc, show.end_time_utc)
         .then(p => {
           show.playlist_executed = p.saved_items;
@@ -197,7 +206,7 @@ function currentShow(req, res) {
         })
         .catch(err => {
           console.log(
-            'error in events > root > currentShow > findOne for playlist',
+            'error in events > root > nextShow > findOne for playlist',
           );
           console.error(err);
           return res.status(500).json(err);
@@ -216,7 +225,7 @@ function currentShow(req, res) {
     .populate(populateMasterEventShowDetails())
     .then(processEventResults)
     .catch(err => {
-      console.log('error in events > root > currentShow');
+      console.log('error in events > root > nextShow');
       console.error(err);
       return res.status(500).json(err);
     });
