--- server/v1/controllers/events/series/index.js
@@ -1,7 +1,0 @@
-const remove = require('./remove');
-const update = require('./update');
-
-module.exports = {
-  remove,
-  update,
-};
